# api-management/message-api/openapi-with-jwt.yaml
swagger: "2.0"
info:
  title: "Go Messages API"
  description: "API REST básica en Go para manejar mensajes"
  version: "1.0.0"
host: senserrich-test-293912.appspot.com # e.g., senserrich-test-293912.appspot.com
x-google-endpoints:
  - name: senserrich-test-293912.appspot.com # e.g., senserrich-test-293912.appspot.com
    allowCors: true
schemes:
  - https
produces:
  - "application/json"
consumes:
  - "application/json"
paths:
  /messages:
    get:
      summary: "Obtener todos los mensajes"
      description: "Lista todos los mensajes"
      operationId: listMessages
      security: # <-- MODIFICADO: Ahora requiere API Key O JWT
        - api_key: []
        - google_jwt: [] # <-- ¡NUEVO! Requiere JWT de Google
      x-google-quota:
        metricCosts:
          list_requests: 1
      responses:
        "200":
          description: "Lista de mensajes"
          schema:
            type: array
            items:
              $ref: "#/definitions/Message"
    post:
      summary: "Crear un nuevo mensaje"
      description: "Crea un nuevo mensaje"
      operationId: createMessage
      security: # <-- MODIFICADO
        - api_key: []
        - google_jwt: [] # <-- ¡NUEVO!
      parameters:
        - name: "body"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/NewMessage"
      responses:
        "201":
          description: "Mensaje creado exitosamente"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Solicitud inválida"

  "/messages/{id}":
    parameters:
      - name: id
        in: path
        required: true
        type: integer
        format: int32
    get:
      summary: "Obtener un mensaje específico por ID"
      description: "Obtiene un mensaje específico"
      operationId: getMessage
      security: # <-- MODIFICADO
        - api_key: []
        - google_jwt: [] # <-- ¡NUEVO!
      responses:
        "200":
          description: "Mensaje encontrado"
          schema:
            $ref: "#/definitions/Message"
        "404":
          description: "Mensaje no encontrado"
    put:
      summary: "Actualizar un mensaje por ID"
      description: "Actualiza un mensaje"
      operationId: updateMessage
      security: # <-- MODIFICADO
        - api_key: []
        - google_jwt: [] # <-- ¡NUEVO!
      parameters:
        - name: "body"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/NewMessage"
      responses:
        "200":
          description: "Mensaje actualizado"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Solicitud inválida"
        "404":
          description: "Mensaje no encontrado"
    delete:
      summary: "Eliminar un mensaje por ID"
      description: "Elimina un mensaje"
      operationId: deleteMessage
      security: # <-- MODIFICADO
        - api_key: []
        - google_jwt: [] # <-- ¡NUEVO!
      responses:
        "204":
          description: "Mensaje eliminado"
        "404":
          description: "Mensaje no encontrado"

  /:
    get:
      summary: "Página HTML de inicio"
      description: "Página HTML de inicio"
      operationId: serveHTML
      produces:
        - text/html
      responses:
        "200":
          description: OK

x-google-backend:
  address: https://go-rest-api-827153761718.europe-west1.run.app # TU URL DE GO-REST-API

securityDefinitions:
  api_key:
    type: "apiKey"
    name: "key"
    in: "query"
  google_jwt: # <-- ¡NUEVO! Definición del esquema JWT
    authorizationUrl: ""
    flow: "implicit"
    type: "oauth2"
    x-google-issuer: "https://accounts.google.com" # Emisor del token (Google)
    x-google-jwks_uri: "https://www.googleapis.com/oauth2/v3/certs" # URI de las claves públicas de Google
    x-google-audiences: "<project-id>.appspot.com" # El público (audience) esperado en el token JWT

x-google-management:
  metrics:
    - name: "list_requests"
      displayName: "Lista de Solicitudes"
      valueType: INT64
      metricKind: DELTA
    - name: "read_requests"
      displayName: "Solicitudes de Lectura"
      valueType: INT64
      metricKind: DELTA
    - name: "write_requests"
      displayName: "Solicitudes de Escritura"
      valueType: INT64
      metricKind: DELTA

  quota:
    limits:
      - name: "list-limit"
        metric: "list_requests"
        unit: "1/min/{project}"
        values:
          STANDARD: 1
      - name: "read-limit"
        metric: "read_requests"
        unit: "1/min/{project}"
        values:
          STANDARD: 3
      - name: "write-limit"
        metric: "write_requests"
        unit: "1/min/{project}"
        values:
          STANDARD: 2

definitions:
  Message:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      content:
        type: "string"
    required:
      - "id"
      - "content"
  NewMessage:
    type: "object"
    properties:
      content:
        type: "string"
    required:
      - "content"